@@ -20,71 +20,37 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 import logging
-from enum import Enum
+import re
 from functools import lru_cache
 from pathlib import Path
 from typing import Any
-from typing import Optional
-from typing import Union
 from typing import cast
 
 from mkdocs.config.base import Config
 from mkdocs.config.defaults import MkDocsConfig
 from mkdocs.utils import meta as meta_parser
 
+from mkdocs_publisher._shared import links
+
 log = logging.getLogger("mkdocs.publisher._shared.mkdocs_utils")
 
 
-class ConfigChoiceEnum(Enum):
-    def __eq__(self, other) -> bool:
-        if isinstance(other, bool) and self.is_bool:
-            return self._str_to_bool(self.name) is other
-        return self.name.lower() == str(other).lower()
-
-    def __hash__(self):
-        return super().__hash__()
-
-    @staticmethod
-    def _str_to_bool(text) -> bool:
-        if text.lower() == "true":
-            return True
-        elif text.lower() == "false":
-            return False
-        raise ValueError(f"'{text}' cannot be converted into bool value")
-
-    @classmethod
-    def _get_enums(cls, enums: list) -> list:
-        enums_list = []
-        for enum in enums:
-            if enum.is_bool:
-                enums_list.extend([enum.name, cls._str_to_bool(enum.name)])
-            else:
-                enums_list.append(enum.name)
-        return enums_list
-
-    @property
-    def name(self) -> str:
-        return ".".join(str(self).split(".")[1:]).lower()
-
-    @property
-    def is_bool(self) -> bool:
-        return self.value[2]
-
-    @classmethod
-    def default(cls) -> Optional[str]:
-        defaults = [f.name for f in cls if f.value[1]]
-        if len(defaults) == 0:
-            return None
-        elif len(defaults) > 1:
-            raise ValueError(f"Multiple defaults specified: {defaults}")
-        return defaults[0]
-
-    @classmethod
-    def choices(cls) -> list:
-        return cls._get_enums(enums=cast(list, cls))
+ADMONITIONS_RE = re.compile(r"\[!\w+]")
+COMMENTS_RE = re.compile(r"<!--(.*?)-->", flags=re.MULTILINE)
+CODE_BLOCK_RE = re.compile(r"`{3}[^>]*`{3}", flags=re.MULTILINE)
+COMMA_RE = re.compile(" ,")
+ENUMERATIONS_RE = re.compile(r"[0-9#]*\.")
+FOOTNOTES_RE = re.compile(r"^\[[^]]*][^(].*", flags=re.MULTILINE)
+FOOTNOTES_REF_RE = re.compile(r"\[[0-9]*]")
+HEADERS_ID_RE = re.compile(r"{#.*}")
+HTML_RE = re.compile(r"</?[^>]*>")
+NEW_LINE_RE = re.compile(r"\n")
+SPACES_RE = re.compile(r" {2,}")
+SPECIAL_CHARS_RE = re.compile(r"[#*`~\-–_^=<>+|/:]")
+TAB_RE = re.compile(r"\t")
 
 
-def get_plugin_config(mkdocs_config: MkDocsConfig, plugin_name: str) -> Union[None, dict[str, Any], Config]:
+def get_plugin_config(mkdocs_config: MkDocsConfig, plugin_name: str) -> None | dict[str, Any] | Config:
     plugins = mkdocs_config.plugins
     if isinstance(plugins, list):
         for plugin in plugins:
@@ -113,3 +79,31 @@ def read_md_file(md_file_path: Path) -> tuple[str, dict[str, Any]]:  # pragma: n
     with md_file_path.open(encoding="utf-8-sig", errors="strict") as md_file:
         # Add empty line at the end of file and read all data
         return meta_parser.get_data(f"{md_file.read()}\n")
+
+
+def _md_any_link_to_text(match: re.Match) -> str:
+    return f'{match.groupdict()["text"]} '
+
+
+def count_words(content: str):
+    """Count words in markdown content.
+
+    Based on: https://github.com/gandreadis/markdown-word-count
+    """
+    content = re.sub(TAB_RE, "    ", content)
+    content = re.sub(NEW_LINE_RE, " ", content)
+    content = re.sub(SPACES_RE, "    ", content)
+    content = re.sub(COMMENTS_RE, "", content)
+    content = re.sub(FOOTNOTES_RE, "", content)
+    content = re.sub(FOOTNOTES_REF_RE, "", content)
+    content = re.sub(CODE_BLOCK_RE, "", content)
+    content = re.sub(HEADERS_ID_RE, "", content)
+    content = re.sub(HTML_RE, "", content)
+    content = re.sub(SPECIAL_CHARS_RE, "", content)
+    content = re.sub(ENUMERATIONS_RE, "", content)
+    content = re.sub(ADMONITIONS_RE, "", content)
+    content = re.sub(SPACES_RE, " ", content)
+    content = re.sub(links.MD_ANY_LINK_RE, _md_any_link_to_text, content)
+    content = re.sub(COMMA_RE, ",", content)
+
+    return len(content.split())
