@@ -21,22 +21,22 @@
 # SOFTWARE.
 
 import logging
+from collections import UserDict
 from pathlib import Path
-from typing import Optional
-from typing import Union
 from typing import cast
 
 from mkdocs.config.defaults import MkDocsConfig
 
 from mkdocs_publisher._shared import mkdocs_utils
 from mkdocs_publisher.blog.config import BlogPluginConfig
+from mkdocs_publisher.meta.config import MetaPluginConfig
 from mkdocs_publisher.obsidian.config import ObsidianPluginConfig
 
 log = logging.getLogger("mkdocs.publisher._shared.publisher_utils")
 
 
-def get_blog_dir(mkdocs_config: MkDocsConfig) -> Optional[Path]:
-    blog_config: Optional[Union[BlogPluginConfig, Path]] = cast(
+def get_blog_dir(mkdocs_config: MkDocsConfig) -> Path | None:
+    blog_config: BlogPluginConfig | Path | None = cast(
         BlogPluginConfig,
         mkdocs_utils.get_plugin_config(
             mkdocs_config=mkdocs_config,
@@ -44,17 +44,19 @@ def get_blog_dir(mkdocs_config: MkDocsConfig) -> Optional[Path]:
         ),
     )
 
-    blog_dir: Optional[Path] = Path(mkdocs_config.docs_dir).joinpath(blog_config.blog_dir) if blog_config else None
+    blog_dir: Path | None = None
+    if blog_config:
+        blog_dir = Path(mkdocs_config.docs_dir).joinpath(blog_config.blog_dir)
 
     return blog_dir
 
 
-def get_obsidian_dirs(mkdocs_config: MkDocsConfig) -> tuple[list[Path], Optional[Path]]:
+def get_obsidian_dirs(mkdocs_config: MkDocsConfig) -> tuple[list[Path], Path | None]:
     ignored_dirs: list[Path] = []
-    attachments_dir: Optional[Path] = None
+    attachments_dir: Path | None = None
     docs_dir = Path(mkdocs_config.docs_dir)
 
-    obsidian_config: Optional[ObsidianPluginConfig] = cast(
+    obsidian_config: ObsidianPluginConfig | None = cast(
         ObsidianPluginConfig,
         mkdocs_utils.get_plugin_config(
             mkdocs_config=mkdocs_config,
@@ -68,3 +70,27 @@ def get_obsidian_dirs(mkdocs_config: MkDocsConfig) -> tuple[list[Path], Optional
         attachments_dir = docs_dir.joinpath(obsidian_config.attachments_dir)
 
     return ignored_dirs, attachments_dir
+
+
+class PublisherFiles(UserDict):
+    def __init__(self):
+        self._on_serve: bool = False
+        self._mkdocs_config: MkDocsConfig | None = None
+        self._meta_plugin_config: MetaPluginConfig | None = None
+
+        super().__init__()
+
+    @property
+    def on_serve(self) -> bool:
+        return self._on_serve
+
+    @on_serve.setter
+    def on_serve(self, on_serve: bool):
+        self._on_serve = on_serve
+
+    def set_configs(self, mkdocs_config: MkDocsConfig, meta_plugin_config: MetaPluginConfig):
+        self._mkdocs_config = mkdocs_config
+        self._meta_plugin_config = meta_plugin_config
+
+    def add_files(self):
+        raise NotImplementedError
