@@ -22,30 +22,26 @@
 
 import logging
 import re
-from collections import UserDict
 from collections.abc import Generator
 from dataclasses import dataclass
 from dataclasses import field
 from functools import cached_property
 from pathlib import Path
 from typing import Any
-from typing import Optional
 from urllib.parse import quote
 
-from mkdocs.config.defaults import MkDocsConfig
 from mkdocs.structure.files import File
 from mkdocs.structure.files import Files
 
 from mkdocs_publisher._shared import links
 from mkdocs_publisher._shared import mkdocs_utils
+from mkdocs_publisher._shared import publisher_utils
 from mkdocs_publisher._shared import templates
-from mkdocs_publisher._shared import urls
-from mkdocs_publisher.meta.config import MetaPluginConfig
 from mkdocs_publisher.meta.config import OverviewChoiceEnum
 from mkdocs_publisher.meta.config import PublishChoiceEnum
 from mkdocs_publisher.meta.config import TitleChoiceEnum
 
-log = logging.getLogger("mkdocs.publisher._shared.meta_files")
+log = logging.getLogger("mkdocs.publisher.meta.meta_files")
 
 
 HEADINGS_RE = re.compile(r"^#+ (?P<title>[^|#\r\n\t\f\v]+)$")
@@ -56,59 +52,45 @@ class MetaFile:
     is_dir: bool
     path: Path
     abs_path: Path = field(repr=False)
-    file: Optional[File] = field(default=None, repr=False)
-    is_draft: Optional[bool] = field(default=None)
+    file: File | None = field(default=None, repr=False)
+    is_draft: bool | None = field(default=None)
     is_hidden: bool = field(default=False)
     is_overview: bool = field(default=False)
-    redirect: Optional[str] = field(default=None)
-    slug: Optional[str] = field(default=None)
-    title: Optional[str] = field(default=None)
-    url: Optional[str] = field(default=None)
+    redirect: str | None = field(default=None)
+    slug: str | None = field(default=None)
+    title: str | None = field(default=None)
+    url: str | None = field(default=None)
 
     @property
     def name(self) -> str:
         return self.path.name
 
     @property
-    def parent(self) -> Optional[Path]:
+    def parent(self) -> Path | None:
         parent = self.path.parent
         if str(parent) == ".":
             parent = None
         return parent
 
 
-class MetaFiles(UserDict):
+class MetaFiles(publisher_utils.PublisherFiles):
     def __init__(self):
-        self._on_serve: bool = False
-        self._mkdocs_config: Optional[MkDocsConfig] = None
-        self._meta_plugin_config: Optional[MetaPluginConfig] = None
         self._hidden_paths: list[Path] = []
+
         super().__init__()
 
     @property
     def dir_meta_file(self) -> str:
         return self._meta_plugin_config.dir_meta_file
 
-    @property
-    def on_serve(self) -> bool:
-        return self._on_serve
-
-    @on_serve.setter
-    def on_serve(self, on_serve: bool):
-        self._on_serve = on_serve
-
-    def add_hidden_path(self, hidden_path: Optional[Path]):
+    def add_hidden_path(self, hidden_path: Path | None):
         if hidden_path is not None:
             self._hidden_paths.append(hidden_path.relative_to(self._mkdocs_config.docs_dir))
-
-    def set_configs(self, mkdocs_config: MkDocsConfig, meta_plugin_config: MetaPluginConfig):
-        self._mkdocs_config = mkdocs_config
-        self._meta_plugin_config = meta_plugin_config
 
     def _get_title(self, meta_file: MetaFile, meta: dict[str, Any], markdown: str):
         """Calculate title for given file"""
 
-        title: Optional[str] = None
+        title: str | None = None
         mode = self._meta_plugin_config.title.mode
         if mode == TitleChoiceEnum.META:
             title = meta.get(self._meta_plugin_config.title.key_name)
@@ -140,7 +122,7 @@ class MetaFiles(UserDict):
     def _get_slug(self, meta_file: MetaFile, meta: dict[str, Any]):
         """Calculate slug for given file"""
 
-        meta_file.slug = urls.create_slug(  # pragma: no cover
+        meta_file.slug = links.create_slug(  # pragma: no cover
             file_name=meta_file.path.stem,
             slug_mode=self._meta_plugin_config.slug.mode,
             slug=meta.get(self._meta_plugin_config.slug.key_name),
@@ -245,6 +227,7 @@ class MetaFiles(UserDict):
         elif publish in PublishChoiceEnum.published():
             meta_file.is_hidden = False
             meta_file.is_draft = False
+
         else:
             meta_file.is_hidden = False
             meta_file.is_draft = True
@@ -272,7 +255,7 @@ class MetaFiles(UserDict):
                 self._get_metadata(meta_file=meta_file, meta_file_path=meta_file_path)
             else:
                 meta_file.title = str(meta_file.path.stem)
-                meta_file.slug = urls.create_slug(
+                meta_file.slug = links.create_slug(
                     file_name=str(meta_file.name),
                     slug_mode=self._meta_plugin_config.slug.mode,
                     slug=meta_file.path.stem,
@@ -282,6 +265,7 @@ class MetaFiles(UserDict):
                 meta_file.is_draft = not self._meta_plugin_config.publish.dir_default
         else:
             self._get_metadata(meta_file=meta_file, meta_file_path=meta_file.abs_path)
+
         super().__setitem__(path, meta_file)
 
     def _drafts(self, files: bool = False, dirs: bool = False) -> dict[str, MetaFile]:
@@ -338,11 +322,13 @@ class MetaFiles(UserDict):
     def hidden(self) -> dict[str, MetaFile]:
         return self._hidden()
 
-    def add_meta_files(self, ignored_dirs: list[Path]):
+    def add_files(self, ignored_dirs: list[Path] | None = None):
         """Iterate over all files and directories in docs directory"""
+        if not ignored_dirs:
+            ignored_dirs = []
 
         for docs_file in sorted(Path(self._mkdocs_config.docs_dir).rglob("*")):
-            meta_file: Optional[MetaFile] = None
+            meta_file: MetaFile | None = None
             is_ignored = any([docs_file.is_relative_to(ignored_dir) for ignored_dir in ignored_dirs])
 
             if not is_ignored and docs_file.is_dir():
@@ -381,7 +367,7 @@ class MetaFiles(UserDict):
 
         # Replace URL parts that have slug defined
         for position, path_part in enumerate(path_parts):
-            meta_file: Optional[MetaFile] = self.get(str(path_part), None)
+            meta_file: MetaFile | None = self.get(str(path_part), None)
             if meta_file is not None:
                 url_parts[position] = str(meta_file.slug)
 
@@ -424,7 +410,7 @@ class MetaFiles(UserDict):
                 self[file.src_path].url = file.url
         return new_files
 
-    def generate_redirect_page(self, file: File) -> Optional[str]:
+    def generate_redirect_page(self, file: File) -> str | None:
         """Generates content of redirect page"""
 
         meta_file: MetaFile = self[file.src_path]
