@@ -25,7 +25,6 @@ import logging
 from datetime import datetime
 from pathlib import Path
 from typing import Literal
-from typing import Optional
 
 from mkdocs.config import Config
 from mkdocs.config.defaults import MkDocsConfig
@@ -48,10 +47,10 @@ class MetaPlugin(BasePlugin[MetaPluginConfig]):
 
     def __init__(self):
         self._on_serve: bool = False
-        self._attachments_dir: Optional[Path] = None
+        self._attachments_dir: Path | None = None
         self._ignored_dirs: list[Path] = []
         self._meta_files: MetaFiles = MetaFiles()
-        self._meta_nav: Optional[MetaNav] = None
+        self._meta_nav: MetaNav | None = None
 
     def on_startup(self, *, command: Literal["build", "gh-deploy", "serve"], dirty: bool) -> None:  # pragma: no cover
         if command == "serve":
@@ -59,10 +58,10 @@ class MetaPlugin(BasePlugin[MetaPluginConfig]):
         self._meta_files.on_serve = self._on_serve
 
     @event_priority(100)  # Run before any other plugins
-    def on_config(self, config: MkDocsConfig) -> Optional[Config]:  # pragma: no cover
+    def on_config(self, config: MkDocsConfig) -> Config | None:  # pragma: no cover
         # Set some default values
         log.info("Read files and directories metadata")
-        blog_dir: Optional[Path] = publisher_utils.get_blog_dir(mkdocs_config=config)
+        blog_dir: Path | None = publisher_utils.get_blog_dir(mkdocs_config=config)
         self._meta_nav = MetaNav(
             meta_files=self._meta_files,
             blog_dir=blog_dir.relative_to(config.docs_dir) if blog_dir else blog_dir,
@@ -70,7 +69,7 @@ class MetaPlugin(BasePlugin[MetaPluginConfig]):
         self._ignored_dirs, self._attachments_dir = publisher_utils.get_obsidian_dirs(mkdocs_config=config)
         self._meta_files.set_configs(mkdocs_config=config, meta_plugin_config=self.config)
         self._meta_files.add_hidden_path(hidden_path=self._attachments_dir)
-        self._meta_files.add_meta_files(ignored_dirs=self._ignored_dirs)
+        self._meta_files.add_files(ignored_dirs=self._ignored_dirs)
 
         log.info(f"Ignored directories: " f"{[str(d.relative_to(config.docs_dir)) for d in self._ignored_dirs]}")
         log.info(f"Draft files and directories: " f"{list(self._meta_files.drafts.keys())}")
@@ -81,16 +80,14 @@ class MetaPlugin(BasePlugin[MetaPluginConfig]):
         return config
 
     @event_priority(-100)
-    def on_files(self, files: Files, *, config: MkDocsConfig) -> Optional[Files]:  # pragma: no cover
+    def on_files(self, files: Files, *, config: MkDocsConfig) -> Files | None:  # pragma: no cover
         new_files = self._meta_files.clean_redirect_files(files=files)
         new_files = self._meta_files.change_files_slug(files=new_files, ignored_dirs=self._ignored_dirs)
         new_files = self._meta_files.clean_draft_files(files=new_files)
 
         return new_files
 
-    def on_nav(
-        self, nav: Navigation, *, config: MkDocsConfig, files: Files
-    ) -> Optional[Navigation]:  # pragma: no cover
+    def on_nav(self, nav: Navigation, *, config: MkDocsConfig, files: Files) -> Navigation | None:  # pragma: no cover
         removal_list = [*self._meta_files.drafts.keys(), *self._meta_files.hidden.keys()]
         log.debug(f"Nav elements to remove: {removal_list}")
         nav.items = self._meta_nav.nav_cleanup(
@@ -120,9 +117,9 @@ class MetaPlugin(BasePlugin[MetaPluginConfig]):
             page.meta["search"] = {"exclude": True}
 
     @event_priority(-100)  # Run after all other plugins
-    def on_post_page(self, output: str, *, page: Page, config: MkDocsConfig) -> Optional[str]:  # pragma: no cover
+    def on_post_page(self, output: str, *, page: Page, config: MkDocsConfig) -> str | None:  # pragma: no cover
         if page.file.src_path in self._meta_files:
-            redirect_page: Optional[str] = self._meta_files.generate_redirect_page(file=page.file)
+            redirect_page: str | None = self._meta_files.generate_redirect_page(file=page.file)
             if redirect_page:
                 output = redirect_page
         return output
