@@ -22,10 +22,9 @@
 
 import logging
 import re
+from collections.abc import Callable
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Callable
-from typing import Optional
 
 import watchdog.events
 from mkdocs.config import Config
@@ -56,7 +55,7 @@ log = logging.getLogger("mkdocs.publisher.obsidian.plugin")
 @dataclass
 class Comment:
     comment: str
-    is_html_comment: Optional[bool] = False
+    is_html_comment: bool | None = False
 
     def __repr__(self):
         return f"<!--{self.comment}-->" if self.is_html_comment else ""
@@ -66,9 +65,9 @@ class ObsidianPlugin(BasePlugin[ObsidianPluginConfig]):
     supports_multiple_instances = False
 
     def __init__(self):
-        self._backlink_links: Optional[BacklinkLinks] = None
+        self._backlink_links: BacklinkLinks | None = None
         self._backlinks: dict[str, list[Link]] = {}
-        self._md_links: Optional[MarkdownLinks] = None
+        self._md_links: MarkdownLinks | None = None
         self._vega_pages: list[Page] = list()
 
     def _normalize_comments(self, match: re.Match) -> str:
@@ -76,12 +75,12 @@ class ObsidianPlugin(BasePlugin[ObsidianPluginConfig]):
         comment.is_html_comment = self.config.comments.inject_as_html
         return str(comment)
 
-    def on_config(self, config: MkDocsConfig) -> Optional[Config]:
+    def on_config(self, config: MkDocsConfig) -> Config | None:
         self._backlink_links = BacklinkLinks(mkdocs_config=config, backlinks=self._backlinks)
         self._md_links = MarkdownLinks(mkdocs_config=config)
         return config
 
-    def on_files(self, files: Files, *, config: MkDocsConfig) -> Optional[Files]:
+    def on_files(self, files: Files, *, config: MkDocsConfig) -> Files | None:
         if self.config.vega.enabled:
             resources.add_extra_css(
                 stylesheet_file_name="obsidian.min.css",
@@ -90,7 +89,7 @@ class ObsidianPlugin(BasePlugin[ObsidianPluginConfig]):
             )
         return files
 
-    def on_nav(self, nav: Navigation, *, config: MkDocsConfig, files: Files) -> Optional[Navigation]:
+    def on_nav(self, nav: Navigation, *, config: MkDocsConfig, files: Files) -> Navigation | None:
         if self.config.backlinks.enabled:
             log.info("Parsing backlinks")
             for file in files:
@@ -102,7 +101,7 @@ class ObsidianPlugin(BasePlugin[ObsidianPluginConfig]):
         return nav
 
     @event_priority(100)  # Run before all other plugins
-    def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files) -> Optional[str]:
+    def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files) -> str | None:
         # TODO: add verification if relative backlinks are enabled in .obsidian config
         markdown = self._md_links.normalize_links(markdown=markdown, current_file_path=Path(page.file.src_uri))
         if self.config.comments.enabled:
@@ -133,7 +132,7 @@ class ObsidianPlugin(BasePlugin[ObsidianPluginConfig]):
                 markdown = f"{markdown}{backlink_render}"
         return markdown
 
-    def on_post_page(self, output: str, *, page: Page, config: MkDocsConfig) -> Optional[str]:
+    def on_post_page(self, output: str, *, page: Page, config: MkDocsConfig) -> str | None:
         if self.config.vega.enabled and page in self._vega_pages:
             # TODO: embed scripts to assets and give possibility to serve from site_dir
             html_modifier = HTMLModifier(markup=output)
@@ -144,9 +143,7 @@ class ObsidianPlugin(BasePlugin[ObsidianPluginConfig]):
 
         return output
 
-    def on_serve(
-        self, server: LiveReloadServer, *, config: MkDocsConfig, builder: Callable
-    ) -> Optional[LiveReloadServer]:
+    def on_serve(self, server: LiveReloadServer, *, config: MkDocsConfig, builder: Callable) -> LiveReloadServer | None:
         server.unwatch(config.docs_dir)
 
         docs_dirs_to_skip = [str(Path(config.docs_dir) / self.config.obsidian_dir)]
