@@ -46,7 +46,7 @@ from mkdocs_publisher.obsidian.plugin import ObsidianPlugin
         ),
         (
             "Lorem ipsum dolor sit [amet](file.md#anchor part), consectetur adipiscing elit.",
-            "Lorem ipsum dolor sit [amet](file.md#anchor part), consectetur adipiscing elit.",
+            "Lorem ipsum dolor sit [amet](file.md#anchor-part), consectetur adipiscing elit.",
         ),
         (
             'Lorem ipsum dolor sit [amet](file.md "title"), consectetur adipiscing elit.',
@@ -54,11 +54,11 @@ from mkdocs_publisher.obsidian.plugin import ObsidianPlugin
         ),
         (
             'Lorem ipsum dolor sit [amet](file.md#anchor part "title"), ' "consectetur adipiscing elit.",
-            'Lorem ipsum dolor sit [amet](file.md#anchor part "title"), ' "consectetur adipiscing elit.",
+            'Lorem ipsum dolor sit [amet](file.md#anchor-part "title"), ' "consectetur adipiscing elit.",
         ),
         (
             "Lorem ipsum dolor sit [amet](#just/an anchor), consectetur adipiscing elit.",
-            "Lorem ipsum dolor sit [amet](#just/an anchor), consectetur adipiscing elit.",
+            "Lorem ipsum dolor sit [amet](#justan-anchor), consectetur adipiscing elit.",
         ),
         (
             "Lorem ipsum dolor sit [[file]], consectetur adipiscing elit.",
@@ -70,7 +70,7 @@ from mkdocs_publisher.obsidian.plugin import ObsidianPlugin
         ),
         (
             "Lorem ipsum dolor sit [[file#anchor part]], consectetur adipiscing elit.",
-            "Lorem ipsum dolor sit [file > anchor part](file.md#anchor part), " "consectetur adipiscing elit.",
+            "Lorem ipsum dolor sit [file > anchor part](file.md#anchor-part), " "consectetur adipiscing elit.",
         ),
         (
             "Lorem ipsum dolor sit [[file|amet]], consectetur adipiscing elit.",
